Reimplemented GAPainter function. Function parameters will be added later. I changed the Painter class to make it a thread. If necessary, I kept the old version as CopyOfPainter.
Known issues: the stop button doesn't work anymore, some changes in the 
program will be needed

Also changed the help messages type: now an image explaining the
function will have to be made.

I finally added a "true" main class, PhotoWork, in the package main. It
calls the "create" method in GUI which is just the previous main method.